Move active_flows from VrouterStatsAgent to VrouterStatsAgent.flow_rate
This is required by UI to chart a graph based on active flow count.
Moving ensures that series of flow counts information is available for UI to
query because new location is under flow_rate which has tags annotation.
active_flows should be interpreted as dial type as aggregate value is sent
always.

Change-Id: I2237cb817a1791f3431e48890fbe0286f184279f
Closes-Bug: #1545654