Fix issue of physical interface field not sent in Vrouter Uve
Problem
------
Agent sends Physical interface field only once at startup. If
physical interface is not yet populated in UVE data-structure
when the agent sends VrouterAgent UVE for the first time, this
field is never sent.

Fix
——
Send this field along with other fields which are sent periodically
if there are any changes.

Change-Id: Id45bd6e2dbaedba5a4f085b686a3c4093a579fb3
Closes-Bug: #1572578