Fixes: contrail-status should indicate ifmap as down if database is empty
In the health-check for IFMAP, adding a check to check whether the
default global system config is there.

If not publish the IFMAP server as Down with the discovery server
with relevant reason and also bring the ConnectionStatus for IFMAP
in API server to down so that the contrail-status reports that
the IFMAP server connection is not functional.

Change-Id: Ia5c1589055f7795ca4a145403c583588df94cec5
Closes-Bug: 1546689