Use adaptive diff stats while exporting flows
Problem
———----
When a flow export message is dropped, the diff bytes/packets
computed for that flow is not retained. Next time when is flow
is being exported the new diff bytes/packets computed will be
since the last time it was attempted to be sent. This will
result in incorrect diff bytes/packets being exported for flows

Fix
--—
When a flow sample is dropped retain the diff bytes/packets for
this flow. Next time when this flow is being exported the diff
bytes/packets computed will include the previous diff bytes/packets
retained for that flow.

Also replace the count "flow-export-msg-drops" with the following
--drops due to disable of flow-export
--drops because of subjecting flows to flow-sampling

Closes-Bug: #1572473
(cherry picked from commit 339a217b0c3223705844f518983d14d012724e9d)

Change-Id: I5990fd688f83004489aca008d41d802d03b4fd29