Keep track of interest and advertised as separate resources.
Both interest and advertised need to be considered as separate resources since
each has a life of its own, they dont move in lockstep and should not be
coupled in the config tracking. Adding separete config tracking abilities for
advertised too.

This is required for race conditions like the following:
1. Node N1 has both interest and advertised set with client C1's index.
2. The graph-walker processes a link-delete which makes node N1 unreachable
for client C1. It removes the interest for C1 from N1 and triggers a
table-change to let the exporter process it. Until the exporter processes it,
update for N1 (to be sent to C1) will not be added to the updateQ.
3. C1 gets a NotReady i.e. its going away. The ifmap-server processes the client
delete and as part of that cleanup goes through the updateQ for any
pending updates to C1. It does not find anything and cleans up the client C1
from all its data-structures. C1 is gone.
4. Exporter finally processes the trigger from the graph-walker from #1 and
adds a delete-update for C1 to the UpdateQ since N1's advertised has C1's index
set. But, C1 is gone and does not need to see this delete-update and nobody will
clean it up. This will also keep the IFMapState for N1 alive with advertised
still set to C1. So, if C1 (or another client C2) comes back, we will not send
N1 to him since the advertised is set.

Change-Id: I32efa5bc5ded55c8e46d1ebe951ee59542ad1587
Closes-Bug: #1551398