Agent to query all available named Servers in the system.
Agent to request all available named servers from Disocvery Server while
connecting to two in the list to send DNS records, but querying all.

Querying all is needed in transitionarty cases where an available
named server is DOWN/crashed. In such a case, we might be sending
queries to a dead named. Discovery server is informed about the
unresponsive named server, but the resubscribe to learn about new
set of DNS server can happen as long as for 30 mins.

At this point not all agent are connected to same two dns-servers
and hence query from all is needed and record updates to any two.

Also we need exclusion for "Agent::ControllerXmpp" task which updates
the response from DiscoveryServer and "Agent::Services" that uses
the list to send queries.

Change-Id: Icd821d799244d1c58da79733ecad9ce09b2fe852
Closes-Bug:1551987