Limit retries on pycassa connection pool
Currently, pycassa connection pool has unlimited retries. As a result, if
cassandra server goes down, this condition is not detected and the function
hangs forever. With this change, we limit the retries to 30. After that, an
exception will be thrown, which can be caught to return appropriate error to the
caller.

Change-Id: Iac071ffa07c1ec09a3118ce12d2bea80cf3e452e
Closes-Bug: 1565681
(cherry picked from commit b627a75b8eacbe32b650a2409d858c774939fecb)