Fix Agent crash in FlowStatsCollector::DeleteEvent
Issue
-----
FlowEntry was pointing to deleted and freed FlowStatsCollector. In this state
when DeleteEvent was received we enqueued Delete Request on stale
FlowStatsCollector which resulted in crash.

Fix
---
When DeleteEvent is received by FlowStatsManager after enqueuing it to
FlowStatsCollector, reset the FlowStatsCollector pointer of FlowEntry. This will
be set when AddEvent is received

Closes-Bug: #1577084
(cherry picked from commit 1e9a732a87954d71703180a7a0e6fd5c49dfb48e)

Change-Id: I7af9a76b58627da3c184c70127f1956c7c64a573