In BGP service if VM sends TTL 1 session doesnt come up.
If TTL is sent as 1 by VM, modify it to 255 and for reverse keep it at 2.
(Logic for 2 is that vrouter decrements TTL because of routing and hence VM
 should get packet back with 1).
If TTL is anything else than 1, then honor it.

Closes-bug: #1567586

Conflicts:
	src/vnsw/agent/pkt/flow_entry.cc
	src/vnsw/agent/pkt/pkt_flow_info.h
	src/vnsw/agent/vrouter/ksync/flowtable_ksync.cc

(cherry picked from commit b7cc0d7cd984d89b382b163d5b636a88814288f5)

Change-Id: I3f9f148c0bf987346fe8096e8d4d72e29a372b9b