Do not send default load balance attribute over xmpp
Receiver (agent) already falls back to default if load-attribute is not sent.
This attribute is long and verbose and not pleasant for eyes. Most of the time,
it would point to the default 5-tuple set. Hence, no need to send it across the
wire.

It would have been nice if schema supported default values for the vector.
Then receiver code would have been even simpler. But it is not the case at the
moment.

Change-Id: Ia5a18f5896e40f1aaee4238b57be538620a529c1
Closes-Bug: 1579169