Send teardown time in FlowLog messages.
If a Flow Log message is exported for a flow, ensure that
flow teardown time is also sent in subsequent exports. If we
encounter a flow-sample with teardown time for a flow which was
never exported, then subject it sampling just like other flows.

Also treat -1 value for flow-export-rate as disabling of sampling.

Change-Id: I9c5805eb0136dd85705fa3eeaee38786c258d635
Closes-Bug: #1578982