Handle recreate of deleted BGPaaS peer prior to destroy
The basic code path to handle this scenario is same as for regular
peers.  However, there's a bug in updating EndpointToBgpPeerList in
in the BgpServer.  A BGPaaS peer correctly gets removed from this
list when it's deleted. However, it gets added back when processing
the new config if the peer has not been destroyed by then. This can
later result in access to freed memory if the source port is reused.

Fix consists of the following changes:

- Return NULL from PeerManager::PeerLocate if there's a deleted peer
- Do not manipulate EndpointToBgpPeerList if PeerLocate returns NULL
- Insert the peer into EndpointToBgpPeerList when it's resurrected

Change-Id: I8f46878dbabcc69079ce85f20588fd766143d70e
Closes-Bug: 1575517