Fix VmSubscribe race condition with deleted nodes.
If CN receives a VmSubscribe with a UUID when the IFMapNode for the VM in
question in marked deleted, the VmSubscribe is ignored. The VmSubscribe request
should not be ignored. Instead, it should be treated as pending and processed
when VM node is resurrected. If the node is truly deleted and wont ever get
revived, the agent should send a vm-unsubscribe since it sent a vm-subscribe.
That should cleanup the pending entry. If the agent dies before he sends a
vm-unsubscribe, client cleanup should also cleanup the pending entry.

Change-Id: Idb22d3a93b4df46e7ab2874edba358432f48de8f
Closes-Bug: #1577463
(cherry picked from commit c7572b0cbcf51e000f4c43ab16f4b792393601f7)