Fix ToR agent crash in KSyncObjectManager::Process
Issue:
------
ToR agent relies on commiting bulk transactions to
OVSDB server, which on completion can end up as a
no-op to OVSDB server, which internally triggers ack
to complete KSync states on the associated entries
in current bulk context. while delete table processing
being happening it can so happen that for the next
entry pointer obtained there is already a delete op
being triggered and added to Bulk context, while
completion of bulk context it can end up completing
the KSync state for the next entry pointer and end
up deleting it while KSyncObjectManager::Process
still assuming to have a valid pointer without holding
reference to it.

Fix:
----
KSyncObjectManager::Process should take reference to
the KSync Entry pointers it is dealing with to assure
sanity.
Replaced usage of unsafe NotifyEvent with SafeNotifyEvent
for OVSDB based triggers.

Closes-Bug: 1572287
Change-Id: I10fd79d176f0cfc22315a20c47f4db40fe956800