Read and Write TXT records based on length only.
TXT records should not be encoded / decoded as a domain name, they are
not name compressed as domain names and do not end with a 0 length. Also
do the same for any other unsupported record type as well so that they
pass the request and response without dropping them.

Conflicts:
	src/dns/bind/bind_util.cc
	src/dns/test/dns_bind_test.cc

Change-Id: I37f85727a89937c6f33026e90029e537c5d90a32
closes-bug: #1583053