Fix Slow withdraw of Logical Switch from OVSDB
Issue:
------
Logical Switch withdraw from OVSDB server due to VxLAN
unavailable is very slow, as DeleteOvs task sends delete
for one dependent entry and then waits for an ack to
trigger delete for next entry.

Fix:
----
keep reference to the next entry for which a delete needs
to be send and use it to trigger deletes for the next
entry after the process yield completes. this allows the
task to proceed without waiting for Delete Acks

Closes-Bug: 1583556
Change-Id: Ie7fd5c1fe696cb1f47023b4ca3c399eb6d2707df