Fix issue of interface becoming inactive even when it was active.
Issue
Agent interface was marked as “os_oper_state = down” even when the state of the interface in OS was Up. This can happen when agent index for interface is reused for a different interface and OS notification for old interface comes after notification for new interface.

Fix
As soon as we see the delete of interface, mark its index as invalid in VnswInterfaceListenerBase so that it cannot be used in updates until Add is seen again. Also when add is seen update the index in VnswInterfaceListenerBase before equeuing notification for resync of OS attributes. Also re-read OS index for interface whenever notification enqueued by VnswInterfaceListenerBase is processed.

Change-Id: I8811edcc89a809c179d860fb6583324b9816fb69
Closes-Bug: #1506997