Limit number of entries visited per task for ageing
Problem:
Flow ageing currently fires a timer every 50msec. As the flow-table size
increases, the number of entries to visit per timer also increases. This can
lead to large latencies on firing of a timer. It is desirable to limit the
latency

Solution:
When timer is fired, ageing algorithm compute number of entries to visit.
Instead of scanning flow table in context of timer, it starts a task and
visits 256 entries for every run of task. Task continuation is used if there
are more entries pending on completion of task.

Change-Id: Icb10bec94d2923442c453c2b4071ea7ab4d9d702
Partial-Bug: #1568126
Partial-Bug: #1572270
Partial-Bug: #1572471