Fix issue of interface becoming inactive even when it was active.
Issue
Agent interface was marked as “os_oper_state = down” even when the state of the interface in OS was Up. This can happen when agent index for interface is reused for a different interface and OS notification for old interface comes after notification for new interface.

Fix
As soon as we see the delete of interface, mark its index as invalid in VnswInterfaceListenerBase so that it cannot be used in updates until Add is seen again. Also when add is seen update the index in VnswInterfaceListenerBase before equeuing notification for resync of OS attributes. Also re-read OS index for interface whenever notification enqueued by VnswInterfaceListenerBase is processed.

Also define tracebuffer for logging VnswIf events. Also increase the size of tracebuffer for Config tracebuffer.

Closes-Bug: #1506997
(cherry picked from commit fd4e9378164b09b79207183d2e3a37c96e486805)

Change-Id: Ia46f308b6cd008c4aa0906a8be1f0fff65280f91