Check for End-of-Config Marker from IfMapServer.
Check for End-Of-Config before writing to contrail-named.conf, this will avoid
repeated write/erase of contrail-named.conf file.

Change-Id: I24f1b4dd234aedf9932f81c90f9c1e6a5cbcbda0
Closes-Bug:1572852

DNS Records lost even after retries

Retries to named are done in a burst and hence does not succeed, adding timer based retry per
entry so the burst is avoided and contrail-named to able to consume records.

Change-Id: I9f23f9d2208fbd8dfd92e0ad737762ed50d762ba
Closes-Bug:1583566