[VNC OpenStack] List VMI with device_id filters
Before schema 1.06, the VMI parent was the VM but now it's owned by the
project and the anchor with the VM is a back reference.
To keep compatibility, the vnc_openstack code was filtering VMIs with
parent_ids and back_ref_ids set to the device_ids. But since the patch [1]
it's not more possible.
That quick fix initially lists VMIs with back_ref_ids set to device_ids and
if result is empty, tries to list VMIs with parents set to device_ids.
Which is not very efficient in the case that no resources corresponding to
the filters (two api calls instead of one).

[1] https://review.opencontrail.org/#/c/19153/

Change-Id: I1363eaa5f89b4df3875f2ac46de40ebc54836ba7
Closes-Bug: #1586012