Fix accessing deleted vdns obect.
When virtual-dns is deleted, all records configured under the vdns
have their vdns pointer reset. In this case, it was not reset as the
record wasnt added in the list under vdns, although the record was
pointing to vdns. This resulted in accessing deleted vdns entry.

In addition, make sure record is made valid when the appropriate
vdns for it is updated.

Change-Id: Id32f029014a46e8fe3efe35380940706708e7708
closes-bug: #1584174
(cherry picked from commit d09203de75444e54f90ea5bf9b83443e29726281)