lbaas launch issue is fixed.
(id: <index>) will be added in the netns name
once the interface is added to the netns by redhat 7.2.
exists() is modified not to check for the exact match

(@if<index>) will be added to the interface name
once it is attached to the netns by some os. The appended string
should not be passed while removing the interface from the netns.
code is added to strip the appended string if it is.

Change-Id: I3b95878c3228639f3ce5c0220604ce016a2e754b
Closes-bug: 1587711