[VNC API server] Replace simplejson parser by ujson
Replacing the simplejson library parser by the utlrajson library [1] to
improve performance [2].
As the ujson does not support all the builtin json api, and especially
specifying encoding/decoding hook, that fix callbacks to simplejson in
that cases. The encoding/decoding hook ujson implementation is under
review actually [3].
As the ujson library is limited to encodes/decodes numbers [4], that fix
translates the uuids mslong and lslong to string and re-convert to double
from the rest API. The uuid continuing to be store as double in
database, but an external script could be pass to transform it to string
when all API server hosts will be upgraded with that patch.

[1] https://github.com/esnme/ultrajson
[2] https://github.com/esnme/ultrajson#versions
[3] https://github.com/esnme/ultrajson/pull/209
[4] https://github.com/esnme/ultrajson/issues/67#issuecomment-11442016

Change-Id: If1cc7b567f737e9a7bd50308656c1e41f173dbd0
Partial-Bug: #1466777
Closes-bug: #1590835