Fixes: API Server Performance Optimization Placeholder bug
Certain common fields like IdPerms and perms2 are not relevant
to be published or published completely to ifmap.

We are still publishing it to IfMAP since they are part of the
schema. Addiing infra to skip publishing fields not relevant
to IfMap completely or partially publish it to IfMap.

For 20k VMs, 20k VMIs, and 60k objects in total, the Init time
stands at 1:28

Change-Id: I84d0047b47e534c17ad92fe40e7baab7b12a5121
Partial-Bug: 1589014