In L3 mode, no L2 routes are advertised.
- mac of VM is not added with local vm port path.
- mac still gets added with mac-vm binding peer, as vrouter needs it for source
check. However it points to discard NH. Export gets blocked because there is no
local vm peer path.
- broadcast subscription for evpn does not happen, fabric mcast will be send.
- Routes sent from CN in evpn are accepted and no fwding mode is applied.
- OVS is not blocked.

Along with these changes, some more optimization are present.
- Removal of multicast NH, as its not used.
- Moved L2 interface NH under common interface config. It gets added if any of
the modes i.e. ipv4/ipv6/l2 is enabled and gets removed when all are disabled.
This is done as fabric multicast uses these NH and that is sent irrespective of
mode.

Change-Id: Ic48f5c48aa4fd89c7431db2caebea3bd7cf05fd0
Closes-bug: #1564724