Advertise Instance-ip with prefix using alloc-unit
Issue: There is a requirement of assigning alloc-block(power of 2)
vm instance-ip. Today we are creating UC routes for instance-ip
with /32 prefix. Now we need to create and advertise routes
for instance-ip with prefix calculated based on alloc-unit present
in IPAM configuration.
Fix: Store alloc-unit configuration in VnIpam structure and use
it while VM Interface is being created. Add UC route accordingly
and RouteExport will automatically exports the route.

Change-Id: I870a52b3bc732b5476bf17f75c46fe5599077d8a
Closes-Bug: #1595671