Prevent route replication loop
Scenario:
With two VRF importing each other and with routing policy attached to both VRF
to edit the BgpAttr of the routes, each DBEntry notification for route replication
would end up deleting and re-adding back the replicated route and notifying the
secondary route. This would cause an endless route replication loop

Solution:
Search for secondary path while replicating the route, should compare the BgpAttr
with original BgpAttr of the secondary path.

Added test to validate the above fix. i.e. time stamp of the replicated path is
not updated on triggering the notification of the primary route.

Change-Id: I3e6b31e8b36efbc091353e828b4f9dfa809cf5ff
Closes-Bug: #1597687