[VNC API server] Fix get resource with specific fields
When a fields are specifying during read resource, the 'id_perms'
property is not fetch and the API code fails to validate the user
visibility on it. That fix also limit the response to requested fields
in certain cases.
That fields also endorse that props and refs were always returned
irrespective of what fields were requested when a resource is read.

Change-Id: I95e90f540d5c860c1b9b217e5a0aee5d61402816
Closes-bug: #1599175