Handle VN change in interface Notification
Remove the assumption that VN of an interface cannot change without first getting a
notification for NULL VN.

Also fix the following issue.
Also when VN changes for a VMI (clubbing of VMI getting NULL VRF and later new VRF
events into a single event) we were not removing EVPN and Bridge routes for old VRF.

Also added UT to verify this scenario.

Change-Id: Ia3c11ad71558a6e9e215bbb6817ccd9f65253a5d
Closes-Bug: #1596792