[config] Synchronously allocate virtual network ID
To improve config node scalability, that patch move the virtual network
ID allocation from the schema to the vnc API server.
To prevent any issue during the upgrade from version where vni is
allocated by the schema to the version where it's allocated by the vnc
API, that patch also sets the vni if it not sets when the vnc API
received a virtual network create notification. That update will need to
be removed for the future release 3.2.

Change-Id: I53098c36315169035e1a9c2b8474b777eb7f362a
Closes-Bug: #1601919