[config] Synchronously allocate security group ID
To improve config node scalability, that patch move the security group
ID allocation from the schema to the vnc API server.
To prevent any issue during the upgrade from version where sg ID is
allocated by the schema to the version where it's allocated by the vnc
API, that patch also sets the sg ID if it not sets when the vnc API
received a virtual network create notification. That update will need to
be removed for the future release 3.2.

Closes-bug: #1602285

Change-Id: I0d7c1d724194aa3b5bd3f57fec5d132978af5699