AapMode change for existing AAP to be honoured.
RootCause: If Allowed-Address Pair Mode is changed from
active-standby to active-active or vice-versa, change is notified
to Agent operDB, but Agent is not handling the change if AAP is
already installed. Only change in ip-address/prefix-len of AAP is
processed if it is already installed.

Fix: Process the change and modify the route to ECMP when AAP
mode is changed to active-active from default.

Change-Id: I5d32ff5878fb51957e074417c907a0677c59e9d8
Closes-Bug: #1591886