Add ability to selectively skip sending xmpp updates
In scaled setups, we want to minimuze cpu consumed by actual data send over
tcp. Instead of sending and dropping right after read at the receiver end,
we can have an option not send data at all.

This was there, but limited as a binary option. With this change, one can filter
destinations to which the updates send must be skipped, based on regex pattern
match as specified in the environment variable XMPP_SKIP_UPDATE_SEND

Change-Id: I078887186e6df2fb7be5b7a6b6b44d7c40a7e6e8
Partial-Bug: 1464016