Walk Discovery Publisher list during service unavailability.
On published services being detected unavailable (via TCP close),
walk the list of published services recieved from Discovery Server
instead of resubscribing. This is particularly useful when
cfg-node is inaccessible during upgrade etc.,

Also the list returned by the Discovery Server is an ordered list
based on priority and not equal priority, hence connects should be
ordered.

Use std::rotate to demote the node in the list to the end, this
is more effecient as it solves iterator node deallocate/allocate.

Add additional tests to test border conditions.

Change-Id: Iba6115201cac619a69fa9c92f21d29eb0d729e19
Closes-Bug:1605412