Agent core in BridgeRouteEntry
Problem:
On deletion of peer request was enqueued to delete path for same.
For multicast route if its not a force delete it needs a data to identify path
related attributes like vxlan(in case of withdraw etc) along with peer and then delete same.
In above request data was not sent and delete used to assert.
This started happening because of usage of dynamic peer. Before dynamic peer
call to delete path was inline which used to set force delete. Now request to
enqueued to delete but force delete is not known.

Solution:
If deletion has come for a deleted peer, always treat it as force delete and
dont check data.

Closes-bug: #1607456
(cherry picked from commit df67f6d50f58cb2001647d5cc32cf56cc305da9f)

Conflicts:
	src/vnsw/agent/oper/bridge_route.cc

Change-Id: I44efbc50ba60db95321fa39cc1411ec36cf2e6ac