Agent crash @ BgpAsAServiceFlowMgmtTree::GetCNIndex
Problem:
Flow is marked as short-flow and rflow is NULL. Update of this flow is called,
which results in extraction of keys is triggered. For Bgpaas control-node index
is extracted by referring to dest-ip in rflow. Here rflow is NULL and agent
crashes.

Solution:
If no rflow is present then dont extract any keys for Bgpaas. There is no info
to conclude on control-node and tree to be selected.
This is done only for add operation.
During deletion cn_index is stored and need not be re-computed from fabric side
flow.

Change-Id: I0df10ee49d3c0913587401c084ed092c34724fe7
Closes-bug: #1607868
(cherry picked from commit ddee34c0a136dfb8dc9dfad8cd7afb8a0afd292c)