[Config node] Avoid to restart all services if API or IF-map dies
Actually all the config node services are restarted if the API or the
IF-MAP service dies.
If the IF-MAP server dies, the API correctly handle it by unpublish
it on discovery and re-sync when it comes up thanks to that patch
I8123deed561c9344b854e65b60b5735a2ba4d19e.
When the API server dies, updates received by other api-servers
is not reflected to local IF-MAP server. By restarting that IF-MAP server,
control nodes attached to it will move to their backup.

Change-Id: I32f402056264edfa9e1ef01ef7130d2fa59b5c68
Closes-Bug: #1608987