Resync ifmap graph when restart or connection loss
1/ When a ifmap server restarting, the api reseed all graph with data
fetched from database.
2/ If the connection between the api and ifmap server gone down, the api
buffered messages in the ifmap sending queue and send them when the
connection comes up again. If the buffered queue was exceeded, api server
purge the buffer and reseed all the ifmap graph from database.

Also fix an issue in the db-loader script to support new keyspace dict
and add an option to not load zookeeper database.

Change-Id: I8123deed561c9344b854e65b60b5735a2ba4d19e
Closes-Bug: #1622101