Proxy flag was not set for service ip.
This happened becaused evpn route route was withdrawn when health check whent
down, hence mac stitching got removed and proxy flag was removed. Fixing the
same. Along with it interface parameters in vrouter and flow key in vmi, is
relying on vmi being active instead of ipv4/ipv6/bridge active as later can be
down because of health check down.

Change-Id: I288c7b92ebf4497725207e53a7eb5b27769569c9
Closes-bug: #1608460