Keep on trying to create VNC API client from analytics API
The gevent that creates the VNC API client was exiting due to
authentication failure exception. Changed code to handle all
exceptions and keep on trying to create the API client. The
node status will show the API connection down in case we are
not able to create the VNC API client.

Change-Id: I518f6ea6c660fd28ef1b66011554ef5ca5281c1a
Closes-Bug: #1611158