Make schema changes to support private AS manipulation
This patch implements the following changes:

1. Add private-as-action to schema and make appropriate changes to the
configuration parsing code.

2. Fix remove/replace behavior when "all" is not configured.  If all is
false, remove/replace private ASes till we encounter the first public AS
instead of removing/replacing just the first private AS.

3. Use the leftmost public AS in the AS path as the initial replacement
value if the RibOut is iBGP.  Replacing with local AS would introduce a
loop. Use local AS as initial replacement value if the RibOut is eBGP.
Replace private ASes before the first public AS with this replacement
value.

4. Replace private ASes seen after the first public AS with the nearest
public AS i.e. the last public AS seen as the AS Path is processed from
left to right.

Change-Id: I2ed83d737e4bedfc6fe82682fd50a378ed2a32f4
Closes-Bug: 1611147